requirements.txt
discord.py-self==2.1.0
python-dotenv>=1.0.0
openai>=1.0.0
python-binance>=1.0.19
httpx>=0.27.0
